Fixed load seek command.
It previously would seek on the current playing video as opposed to the one loaded. It now will seek the loaded file specified.

Currently, the load command will queue all videos sent to it, thus having the added benifit of being able to specify many videos to load. Will need to confirm if this is desired behaviour.