MPV is a bust. While setting 'rpi' in the config file would suggest Raspberry Pi acceleration, this is not the case, and CPU usuage was high (only noticed when decoding high def video, where frames where dropped). When setting back to 'mmal' (actual hardware decoding), seamless looping and loading broke, as well as my heart.
Upon spending ages trying to workaround this, as well as writing many test scripts for MPV and VLC, I went back to OMX, and found a cheeky little trick to get seamless playing possible.

As frustrating as it is to go full circle back to OMX, from all my other experimentations with other players, it is the most reliable and best performing, and also has the ability to slightly modify the player as needed.

So this branch is a dev branch to reintigrate OMX back into the project, and then back into master.