Fixed issue with network processes not being able to connect if there was no ethernet connection.
If the ethernet was not plugged in, the ethernet interface wouldn't start (?) and processes that directly reference the IP would fail and couldn't connect.
If we change it to localhost, it will always be able to connect. This doesn't change functionality. Once ethernet is connected, it will work normally.

I'm kinda guessing here not gonna lie but it works so whatever.